Fix performances of `argsSettings`
This fixes a performance bug caught by @rampion

The old implementation was not actually caching the values, which this
fixes.

You might wonder why this still uses `unsafeInterleaveIO`, which is
mainly to reduce the amount of time that any given use of
`argsSettings` needs to hold the lock on the `cache` `MVar`.